Replace lifetime events with lifespan async context manager
