+++ src/bartender/web/application.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI
 from fastapi.responses import UJSONResponse
 
 from bartender.web.api.router import api_router
-from bartender.web.lifetime import register_shutdown_event, register_startup_event
+from bartender.web.lifespan import lifespan
 from bartender.web.users.router import include_users_routes
 
 
@@ -24,11 +24,8 @@ def get_app() -> FastAPI:
         redoc_url="/api/redoc",
         openapi_url="/api/openapi.json",
         default_response_class=UJSONResponse,
+        lifespan=lifespan,
     )
-
-    # Adds startup and shutdown events.
-    register_startup_event(app)
-    register_shutdown_event(app)
 
     # Main router for the API.
     app.include_router(router=api_router, prefix="/api")
